+++ desktop/desktop.config.ts
@@ -2,6 +2,6 @@ import { defineDesktopConfig } from '@owdproject/core'
 
 export default defineDesktopConfig({
   theme: '@owdproject/theme-nova',
-  apps: ['@owdproject/app-about'],
+  apps: ['@owdproject/app-about', '@owdproject/app-terminal'],
   modules: [],
 })