+++ packages/core/index.ts
@@ -3,3 +3,7 @@ export {
   defineDesktopConfig,
 } from './runtime/utils/utilDesktop'
 export { registerTailwindPath } from './runtime/utils/utilApp'
+export { defineDesktopModule } from './runtime/utils/defineDesktopModule'
+export { defineDesktopTheme } from './runtime/utils/defineDesktopTheme'
+export { setDesktopExtensionConfig } from './runtime/utils/setDesktopExtensionConfig'
+export { mergeDesktopExtensionConfig } from './runtime/utils/mergeDesktopExtensionConfig'