+++ packages/kit-theme/vitest.config.ts
@@ -1,0 +1,16 @@
+/// <reference types="vitest" />
+import { defineConfig } from 'vite'
+import vue from '@vitejs/plugin-vue'
+
+export default defineConfig(() => ({
+  root: __dirname,
+  cacheDir: '../../node_modules/.vite/packages/kit-theme',
+  plugins: [vue()],
+  test: {
+    watch: false,
+    globals: true,
+    environment: 'jsdom',
+    include: ['test/**/*.{test,spec}.{js,mjs,cjs,ts,mts,cts}'],
+    reporters: ['default'],
+  },
+}))